Add navigation component
