+++ unit-15/jn-svelte-app/src/Increment.svelte
@@ -1,5 +1,5 @@
 <script>
-    import { link } from "svelte-spa-router";
+    import MainNavigation from "./components/MainNavigation.svelte";
     import { count } from "./store.js";
 
     let title = "Increment";
@@ -14,19 +14,13 @@
     <header>
 		<h1><b>Unit-15</b><br>Svelte App</h1>
         <h2>This is {title} page!</h2>
-        <nav>
-            <a href="/" use:link>Home</a>
-            <a href="/increment" use:link>Increment (Count Clicks)</a>
-            <a href="/posts" use:link>Fetch API Posts</a>
-        </nav>
+        <MainNavigation/>
 	</header>
 
 	<section>
 		<p>Count: <b>{$count}</b></p>
 		<button on:click={increment}>Click to Increment</button>
 	</section>
-	<!-- Use fetch component (Fetch.svelte file) -->
-	<!-- <Fetch /> -->
 
     <br>
     <footer>
@@ -44,13 +38,6 @@
 		max-width: 650px;
 		margin: 0 auto;
 	}
-    nav {
-        padding-bottom: 50px;
-        border-bottom: 1px solid lightgray;
-    }
-    nav a:not(:last-child) {
-        margin-right: 6px;
-    }
     section {
         padding: 30px 0;
     }
